provider: Make the ITimeGraphState#getValue return an int
It returned a long which always needed to be cast to an int.

Reviewed-on: https://git.eclipse.org/r/117819
Tested-by: CI Bot
Tested-by: Genevieve Bastien <gbastien+lttng@versatic.net>

+++ analysis/org.eclipse.tracecompass.analysis.os.linux.core/src/org/eclipse/tracecompass/internal/analysis/os/linux/core/resourcesstatus/ResourcesStatusDataProvider.java
@@ -357,7 +357,7 @@ public class ResourcesStatusDataProvider extends AbstractTimeGraphDataProvider<@
                 } else if ((status instanceof Long) && (type == Type.FREQUENCY)) {
                     long s = (long) status;
                     // The value needs to fit in an integer
-                    eventList.add(new TimeGraphState(startTime, duration, s/FREQUENCY_MULTIPLIER, String.valueOf(FREQUENCY_FORMATTER.format(s))));
+                    eventList.add(new TimeGraphState(startTime, duration, (int) (s / FREQUENCY_MULTIPLIER), String.valueOf(FREQUENCY_FORMATTER.format(s))));
                 } else {
                     eventList.add(new TimeGraphState(startTime, duration, Integer.MIN_VALUE));
                 }
